+++ CHANGELOG.md
@@ -6,10 +6,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [[v2.4.0](https://github.com/multiversx/mx-sdk-dapp/pull/604)] - 2023-02-03
+- [Add support for webview provider automatic login](https://github.com/multiversx/mx-sdk-dapp/pull/609)
+
+## [[v2.3.5](https://github.com/multiversx/mx-sdk-dapp/pull/604)] - 2023-02-02
+- [Skip login when address is prefilled](https://github.com/multiversx/mx-sdk-dapp/pull/607)
+
+## [[v2.3.4](https://github.com/multiversx/mx-sdk-dapp/pull/604)] - 2023-02-01
+- [Sanitize web wallet login callback URL](https://github.com/multiversx/mx-sdk-dapp/pull/605)
+
+## [[v2.3.3](https://github.com/multiversx/mx-sdk-dapp/pull/604)] - 2023-01-31
+- [Fixed erronated publish 2.3.2](https://github.com/multiversx/mx-sdk-dapp/pull/604)
+## [[v2.3.2](https://github.com/multiversx/mx-sdk-dapp/pull/603)] - 2023-01-31
+- [Extracted ledger address screens control to separate hook](https://github.com/multiversx/mx-sdk-dapp/pull/602)
+
+
 ## [[v2.3.1](https://github.com/multiversx/mx-sdk-dapp/pull/599)] - 2023-01-26
-    
+
 - [Added support to not show meridiem by default in `getHumanReadableTimeFormat` util function](https://github.com/multiversx/mx-sdk-dapp/pull/597)
 - [Fixed web wallet cancelled login](https://github.com/multiversx/mx-sdk-dapp/pull/598)
+
+- [Add support for webview provider automatic login](https://github.com/multiversx/mx-sdk-dapp/pull/609)
 
 ## [[v2.3.0](https://github.com/multiversx/mx-sdk-dapp/pull/596)] - 2023-01-20